+++ Tests/Private/Get-TVADCGuiConfigurationGroup.Tests.ps1
@@ -1,5 +1,0 @@
-﻿Describe 'Get-TVADCGuiConfigurationGroup' {
-    It 'defines the function' {
-        Get-Content -Path "$PSScriptRoot\..\..\Cmdlets\Private\Get-TVADCGuiConfigurationGroup.ps1" -Raw | Should -Match 'function Get-TVADCGuiConfigurationGroup'
-    }
-}