+++ Tests/Private/New-TVADCGuiConfigurationData.Tests.ps1
@@ -1,5 +1,0 @@
-﻿Describe 'New-TVADCGuiConfigurationData' {
-    It 'defines the function' {
-        Get-Content -Path "$PSScriptRoot\..\..\Cmdlets\Private\New-TVADCGuiConfigurationData.ps1" -Raw | Should -Match 'function New-TVADCGuiConfigurationData'
-    }
-}