+++ Tests/Private/Register-TVADCGuiConfigurationHandler.Tests.ps1
@@ -1,5 +1,0 @@
-﻿Describe 'Register-TVADCGuiConfigurationHandler' {
-    It 'defines the function' {
-        Get-Content -Path "$PSScriptRoot\..\..\Cmdlets\Private\Register-TVADCGuiConfigurationHandler.ps1" -Raw | Should -Match 'function Register-TVADCGuiConfigurationHandler'
-    }
-}