+++ Tests/Private/Register-TVADCGuiGroupFiltering.Tests.ps1
@@ -1,5 +1,0 @@
-﻿Describe 'Register-TVADCGuiGroupFiltering' {
-    It 'defines the function' {
-        Get-Content -Path "$PSScriptRoot\..\..\Cmdlets\Private\Register-TVADCGuiGroupFiltering.ps1" -Raw | Should -Match 'function Register-TVADCGuiGroupFiltering'
-    }
-}